+++ db/migrate/20250627130646_create_admin_tenants.rb
@@ -1,0 +1,16 @@
+class CreateAdminTenants < ActiveRecord::Migration[8.1]
+  def change
+    create_table :admin_tenants do |t|
+      t.string :name
+      t.string :url
+      t.string :login
+      t.string :password
+      t.string :guest_max
+      t.string :guest_rx
+      t.string :guest_tx
+      t.boolean :active
+
+      t.timestamps
+    end
+  end
+end