+++ db/reader_schema.rb
@@ -10,7 +10,20 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2025_06_12_081028) do
+ActiveRecord::Schema[8.1].define(version: 2025_06_27_130646) do
+  create_table "admin_tenants", force: :cascade do |t|
+    t.boolean "active"
+    t.datetime "created_at", null: false
+    t.string "guest_max"
+    t.string "guest_rx"
+    t.string "guest_tx"
+    t.string "login"
+    t.string "name"
+    t.string "password"
+    t.datetime "updated_at", null: false
+    t.string "url"
+  end
+
   create_table "sites", force: :cascade do |t|
     t.string "api_key"
     t.string "controller_id"