+++ app/models/admin/tenant.rb
@@ -1,2 +1,0 @@
-class Admin::Tenant < ApplicationRecord
-end