+++ db/reader_schema.rb
@@ -11,7 +11,17 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema[8.1].define(version: 2025_06_27_130646) do
-  create_table "admin_tenants", force: :cascade do |t|
+  create_table "sites", force: :cascade do |t|
+    t.string "api_key"
+    t.string "controller_id"
+    t.string "controller_url"
+    t.datetime "created_at", null: false
+    t.string "ssid"
+    t.datetime "updated_at", null: false
+    t.string "url"
+  end
+
+  create_table "tenants", force: :cascade do |t|
     t.boolean "active"
     t.datetime "created_at", null: false
     t.string "guest_max"
@@ -20,16 +30,6 @@ ActiveRecord::Schema[8.1].define(version: 2025_06_27_130646) do
     t.string "login"
     t.string "name"
     t.string "password"
-    t.datetime "updated_at", null: false
-    t.string "url"
-  end
-
-  create_table "sites", force: :cascade do |t|
-    t.string "api_key"
-    t.string "controller_id"
-    t.string "controller_url"
-    t.datetime "created_at", null: false
-    t.string "ssid"
     t.datetime "updated_at", null: false
     t.string "url"
   end