+++ test/models/tenant_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class Admin::TenantTest < ActiveSupport::TestCase
+class TenantTest < ActiveSupport::TestCase
   # test "the truth" do
   #   assert true
   # end