+++ app/models/tenant.rb
@@ -1,2 +1,3 @@
 class Tenant < ApplicationRecord
+  has_many :sites, dependent: :destroy
 end