+++ db/migrate/20250630085333_change_tenant.rb
@@ -1,0 +1,8 @@
+class ChangeTenant < ActiveRecord::Migration[8.1]
+  def change
+    remove_column :tenants, :guest_max, :integer
+    remove_column :tenants, :guest_rx, :integer
+    remove_column :tenants, :guest_tx, :integer
+    remove_column :tenants, :url, :string
+  end
+end