+++ db/schema.rb
@@ -10,27 +10,31 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2025_06_27_130646) do
+ActiveRecord::Schema[8.1].define(version: 2025_06_30_090827) do
   create_table "sites", force: :cascade do |t|
+    t.boolean "active"
     t.string "api_key"
-    t.string "controller_id"
     t.string "controller_url"
     t.datetime "created_at", null: false
+    t.integer "guest_max", default: 0, null: false
+    t.integer "guest_rx", default: 0, null: false
+    t.integer "guest_tx", default: 0, null: false
+    t.string "name"
     t.string "ssid"
+    t.integer "tenant_id", null: false
     t.datetime "updated_at", null: false
     t.string "url"
+    t.index ["tenant_id"], name: "index_sites_on_tenant_id"
   end
 
   create_table "tenants", force: :cascade do |t|
     t.boolean "active"
     t.datetime "created_at", null: false
-    t.string "guest_max"
-    t.string "guest_rx"
-    t.string "guest_tx"
     t.string "login"
     t.string "name"
     t.string "password"
     t.datetime "updated_at", null: false
-    t.string "url"
   end
+
+  add_foreign_key "sites", "tenants"
 end